+++ packages/builder/src/builderStore/store/screenTemplates/emptyNewRecordScreen.js
@@ -6,7 +6,7 @@ export default {
 const createScreen = () => ({
   props: {
     _id: "",
-    _component: "@budibase/standard-components/recorddetail",
+    _component: "@budibase/standard-components/newrow",
     _styles: {
       normal: {},
       hover: {},
